Last adjustments
Move some classes and functions in utils.
+ small updates
