+++ nanga_ad_library/utils/__init__.py
@@ -1,5 +1,7 @@
 # nanga_ad_library/utils/__init__.py
 # import classes and methods from the package as a whole
 
+from nanga_ad_library.utils.object_parser import ObjectParser
 from nanga_ad_library.utils.param_checker import check_param_value, check_param_type, enforce_date_param_format
+from nanga_ad_library.utils.request_handler import PlatformResponse, HttpMethod, json_encode_top_level_param
 from nanga_ad_library.utils.version import compare_version_to_default, get_default_api_version, get_sdk_version
